Added initial support for 26.04
It will install the older hyprland from repo until it's released
Then the PPA will be updated to support 26.04

 On branch ubuntu-install
 Your branch is up to date with 'origin/ubuntu-install'.

 Changes to be committed:
	modified:   README.md
	modified:   install-scripts/00-dependencies.sh
	modified:   install-scripts/01-hypr-pkgs.sh
	modified:   install-scripts/hyprland-ppa.sh
	modified:   install.sh

+++ install-scripts/00-dependencies.sh
@@ -126,6 +126,101 @@ if [ "$UBUNTU_VERSION_ID" = "24.04" ]; then
         xwayland
         bc
     )
+elif [ "$UBUNTU_VERSION_ID" = "26.04" ]; then
+    dependencies=(
+        build-essential
+        cmake
+        cmake-extras
+        curl
+        findutils
+        gawk
+        gettext
+        git
+        glslang-tools
+        gobject-introspection
+        golang
+        hwdata
+        jq
+        pkg-config
+        libmpdclient-dev
+        libnl-3-dev
+        libasound2-dev
+        libstartup-notification0-dev
+        libwayland-client++1
+        libwayland-dev
+        libcairo-5c-dev
+        libcairo2-dev
+        libsdbus-c++-bin
+        libegl-dev
+        libegl1-mesa-dev
+        libdrm-dev
+        libgbm-dev
+        libinput-dev
+        libudev-dev
+        libseat-dev
+        libdisplay-info-dev
+        libliftoff-dev
+        libpixman-1-dev
+        libtomlplusplus-dev
+        libpango1.0-dev
+        libgdk-pixbuf-2.0-dev
+        libxcb-keysyms1-dev
+        libwayland-client0
+        libxcb-ewmh-dev
+        libxcb-cursor-dev
+        libxcb-icccm4-dev
+        libxcb-composite0-dev
+        libxcb-res0-dev
+        libxcb-xfixes0-dev
+        libxcb-render0-dev
+        libxcb-randr0-dev
+        libxcb-render-util0-dev
+        libxcb-util-dev
+        libxcb-xkb-dev
+        libxcb-xinerama0-dev
+        libxkbcommon-dev
+        libxkbcommon-x11-dev
+        libxcursor-dev
+        meson
+        ninja-build
+        nm-tray
+        openssl
+        psmisc
+        python3-mako
+        python3-markdown
+        python3-markupsafe
+        python3-yaml
+        python3-pyquery
+        qt6-base-dev
+        qt6-base-private-dev
+        qt6-wayland-dev
+        qt6-wayland-private-dev
+        qt6-declarative-dev
+        qml6-module-qtcore
+        qml6-module-qtquick-layouts
+        qt6-tools-dev
+        qt6-tools-dev-tools
+        rsync
+        spirv-tools
+        unzip
+        vulkan-validationlayers
+        vulkan-utility-libraries-dev
+        wayland-protocols
+        xdg-desktop-portal
+        xwayland
+    )
+
+    # hyprland dependencies (runtime libs only; do NOT install hyprcursor-util or libhyprcursor-dev to avoid conflicts with PPA's libhyprcursor1)
+    hyprland_dep=(
+        bc
+        binutils
+        libc6
+        libcairo2
+        libdisplay-info2
+        libdrm2
+        hyprland-guiutils
+        libpam0g-dev
+    )
 else
     dependencies=(
         build-essential
@@ -255,8 +350,8 @@ for PKG in "${conflicts[@]}"; do
     uninstall_package "$PKG" 2>&1 | tee -a "$LOG" || true
 done
 
-# Proactively remove conflicting Ubuntu package if present (overlaps /usr/bin/hyprcursor-util)
-if dpkg -l | grep -q '^ii  hyprcursor-util '; then
+# Proactively remove conflicting Ubuntu package only when using PPA (overlaps /usr/bin/hyprcursor-util)
+if [ "${HYPR_USE_PPA:-0}" = "1" ] && dpkg -l | grep -q '^ii  hyprcursor-util '; then
     echo "${INFO} Removing conflicting hyprcursor-util (Ubuntu repo) to allow libhyprcursor1 from PPA" | tee -a "$LOG"
     sudo apt -y purge hyprcursor-util 2>&1 | tee -a "$LOG" || true
     sudo apt -y autoremove 2>&1 | tee -a "$LOG" || true
